Fix type names starting with lowercase 'a'
Type names that start with a lowercase 'a' cause an error when used:

$ typeset -T al=(typeset bar)
$ al foo=(bar=testset)
/usr/bin/ksh: al: : invalid variable name

The error occurs because when the parser checks for the alias builtin
(to set 'assignment' to two instead of one), only the first letter of
'argp->argval' is checked (rather than the entire string). This was
fixed in ksh93v- by comparing argp->argval against alias, but in ksh93u+m
the check can simply be removed because it is only run when a builtin has
the BLT_DCL flag. The alias builtin is never assigned the BLT_DCL flag.

src/cmd/ksh93/sh/parse.c:
- Remove the bugged (and otherwise unused) check for the alias builtin.

src/cmd/ksh93/tests/types.sh:
- Add a regression test for type names starting with a lowercase 'a'.