Fix leaks by moving sh_iorestore() call instead
Moving it to where the command's exit status gets set seems to fix
all the leaks without the need toa add new calls or any
special-casing.

I thought longjmp might kill this, but it doesn't seem to. Even
when triggering an error in a special builtin (such as 'typeset')
with a process substitution, I don't get a leak.

I traced the sh_iorestore calls by getting them to output the line
and file they were called from, and it turns out that for longjmps,
there is an sh_iorestore() call in main.c, line 418.