edit/edit.c: Fix how the bell character is handled after SIGWINCH
To set a window title in bash and zsh, the $PS1 prompt can be set with
the title placed between $'\E]0;' and $'\a':

set -o emacs    # Or vi mode
typeset -A fmt=(
	[start_title]=$'\E]0;'
	[end_title]=$'\a'
)
PS1="${fmt[start_title]}$(hostname): $(uname)${fmt[end_title]}\$ "

This also works in ksh unless the shell receives SIGWINCH. With a $PS1
that sets a window title, the prompt breaks until two interrupts are
received. This is caused by ed_setup() skipping $'\a' (the bell
character), so the fix is to avoid skipping it when handling SIGWINCH.
For this fix to have any effect, shp->winch is set back to zero after
sfpkrd().