Fix the <>; redirection operator in -c scripts
This bug was first reported at https://github.com/att/ast/issues/9.

The <>; operator doesn't work correctly if it's used as the last command
of a -c script. Reproducer:
  $ echo test > a; ksh -c 'echo x 1<>; a'; cat a
  x
  st
This bug is caused by ksh running the last command of -c scripts with
execve(2) instead of posix_spawn(3) or fork(2). The <>; operator is
noted by the man page as being incompatible with the exec builtin (see
also the ksh93u+ man page), so it's not surprising this bug occurs
when ksh runs a command using execve:

> <>;word cannot be used with the exec and redirect built-ins.

The ksh2020 fix simply removed the code required for ksh to use this
optimization at all. It's not a performance friendly fix and only papers
over the bug, so this commit provides a better fix.

src/cmd/ksh93/sh/xec.c:
- The IOREWRITE flag is set when handling the <>; operator, so to fix
  this bug avoid exec'ing the last command if it uses <>;. See also
  commit 17ebfbf6, which fixed another issue related to the execve
  optimization.

src/cmd/ksh93/tests/io.sh:
- Enable a regression test that was failing because of this bug.
- Add the reproducer from https://github.com/att/ast/issues/9 as a
  regression test.

src/cmd/ksh93/tests/options.sh:
- This bugfix was causing the options regression test to segfault when
  run under shcomp. The cause is the same as https://github.com/ksh93/ksh/issues/165,
  so as a workaround avoid parsing the command substitution with shcomp.
  This workaround should also avoid the other problem detailed in
  https://github.com/ksh93/ksh/issues/274.