Fix $LINENO after running a virtual subshell
There is a TODO note in variables.sh that notes the value of $LINENO is
wrong after a virtual subshell. The following script should print '6',
but the bug causes it to print '1' instead:
  $ cat /tmp/lineno
  #!/bin/ksh
  (
      unset LINENO
      :
  )
  echo $LINENO

This bug started to occur after the bugfix applied in 7b994b6a. However,
that commit is not where the cause of bug was (when that bugfix is applied
to ksh versions 2008-07-25 through 2012-01-01, $LINENO works fine).
Rather, the cause of this bug was introduced in 93u+ 2012-02-29. In that
version, the mp->nvfun pointer was changed to only point to np->nvfun if
the variable can be freed from memory. This is what caused 7b994b6a to
break $LINENO in subshells, so to fix this bug the mp->nvfun pointer must
point to np->nvfun, even when the variable isn't freed from memory.

src/cmd/ksh93/sh/subshell.c: nv_restore():
- Always point mp->nvfun to np->nvfun. To prevent crashes, the value of
  np->nvfun->nofree is set to the value given by the nofree variable,
  which is set before _nv_unset. See also commit 7e7f1372, which fixed
  a crash that happened because _nv_unset discards the NV_NOFREE flag.

src/cmd/ksh93/tests/variables.sh:
- Remove the workaround for $LINENO after a virtual subshell.
- Add a regression test for the value of $LINENO when unset in a virtual
  subshell, then used after the subshell. Note that before commit
  997ad43b $LINENO's value was corrupted after being unset in a
  subshell, so the test checks for corruption of the $LINENO variable
  (in prior commits $LINENO was set to '49' because of the previous
  bug).