Set a different pseudorandom seed after forking a subshell
This fixes the issue detailed in bug #285. The main problem is that ksh
doesn't set a new pseudorandom seed after forking a subshell. The
reproducer uses a redirect to fork a subshell, although the redirect can
be replaced with 'ulimit -t unlimited' to achieve the same result.
When ksh forks a subshell, it should not continue to use the same
pseudorandom seed as the subshell. This is because while nget_rand() can
get a different random number using the same seed, it only does so if
it knows what the last random number was. Once a subshell forks,
nget_rand() (in the parent shell) isn't able to know what the last
random number was in the forked subshell.

src/cmd/ksh93/sh/subshell.c: sh_subfork():
- After forking a subshell, set a pseudorandom seed in the parent shell
  that is different from the one the forked subshell is using. This seed
  is based off of the last obtained random number. If $RANDOM is unset,
  a new seed will not be set (since there is no random number to use).

src/cmd/ksh93/tests/variables.sh
- Add two regression tests for $RANDOM after it's used in a forked
  subshell. The second test makes sure we are actually using a
  different pseudorandom seed, rather than the next number from the
  same seed.
- Add a regression test for a possible crash when $RANDOM is unset
  before launching a forked subshell.

Fixes: https://github.com/ksh93/ksh/issues/285