When invoking a non-interpreter (hash-bang) script, there is a fork, but there is no exec syscall in the child.  The existing command line is overwritten in fixargs() with the name of the new script and associated arguments.  If the new command line is longer than the existing one, it is truncated.  This works well when calling a script with a shorter name.
However, it generates a misleading name in the common scenario where a
script is invoked from an interactive shell.  For instance, if
"/tmp/script" is invoked, "ksh" gets replaced with "/tm" in "ps"
output.  (Leaving the original command in place like bash does would
probably be preferable to this).

This change scavenges space from the environ section that follows the
command-line args in the loaded ELF layout and presents the entire
command (or as much as fits in that section.)  On Linux, this will
overwrite the beginning of /proc/PID/environ.  This cannot be helped,
but far more users consult the output of "ps" to view the command line
than those looking at the proc-environ file.  The actual environment
of the process is not changed in any case.  Overall, the behavior is
somewhat similar to that in the AST fork.

The code that clears remants of the prior command works as before and
clears the buffer only up to the start of the environ section.

This has been tested on Linux and Darwin, but not on other platforms.